Move all component in separate file for resource management
