+++ frontend/packages/app/src/app/pages/resource-management/components/empty/emptyTableCell.tsx
@@ -1,0 +1,52 @@
+/**
+ * External dependencies.
+ */
+import { useContext } from "react";
+import { ResourceTableCell, EmptyTableCell as ResourceEmptyTableCell } from "@next-pms/resource-management/components";
+
+/**
+ * Internal dependencies.
+ */
+import type { EmptyTableCellProps } from "./types";
+import { ResourceFormContext } from "../../store/resourceFormContext";
+
+/**
+ * This component is responsible for rendering the empty table cell.
+ *
+ * @param props.cellClassName class of the given cell.
+ * @param props.textClassName class of the given cell text.
+ * @param props.title title of the given cell.
+ * @param props.onCellClick on cell click event for cell based on this will decide where to show + icon on hover or not.
+ * @returns React.FC
+ */
+const EmptyTableCell = ({ cellClassName, title, textClassName, onCellClick }: EmptyTableCellProps) => {
+  const { permission: resourceAllocationPermission } = useContext(ResourceFormContext);
+
+  if (!onCellClick || !resourceAllocationPermission.write) {
+    return (
+      <ResourceTableCell
+        type="default"
+        cellClassName={cellClassName}
+        cellTypographyClassName={textClassName}
+        value="-"
+      />
+    );
+  }
+
+  if (onCellClick) {
+    return (
+      <ResourceEmptyTableCell
+        cellClassName={cellClassName}
+        title={title}
+        textClassName={textClassName}
+        onCellClick={onCellClick}
+      />
+    );
+  }
+
+  return (
+    <ResourceTableCell type="default" cellClassName={cellClassName} cellTypographyClassName={textClassName} value="-" />
+  );
+};
+
+export { EmptyTableCell };
