+++ frontend/packages/app/src/app/pages/resource-management/components/empty/types.ts
@@ -1,0 +1,8 @@
+interface EmptyTableCellProps {
+  cellClassName?: string;
+  textClassName?: string;
+  title?: string;
+  onCellClick?: () => void;
+}
+
+export type { EmptyTableCellProps };
