+++ frontend/packages/app/src/app/pages/resource-management/components/resource-allocation-list/index.ts
@@ -1,0 +1,3 @@
+export { DeleteIcon } from "./deleteIcon";
+export { ResourceAllocationList } from "./resourceAllocationList";
+export { ResourceAllocationCard } from "./resourceAllocationCard";
