+++ frontend/packages/app/src/app/pages/resource-management/project/components/expand-view/index.ts
@@ -1,0 +1,2 @@
+export { ResourceExpandView } from "./expandView";
+export { ExpandViewCell } from "./expandViewCell";
