+++ frontend/packages/app/src/app/pages/resource-management/project/components/table/index.ts
@@ -1,0 +1,3 @@
+export { ResourceProjectTable } from "./resourceProjectTable";
+export { ResourceProjectTableBody } from "./resourceProjectTableBody";
+export { ResourceProjectTableCell } from "./resourceProjectTableCell";
