+++ frontend/packages/app/src/app/pages/resource-management/team/components/expand-view/index.ts
@@ -1,0 +1,2 @@
+export { ExpandViewCell } from "./expandViewCell";
+export { ResourceExpandView } from "./resourceExpandView";
