+++ frontend/packages/app/src/app/pages/resource-management/team/components/table/index.ts
@@ -1,0 +1,3 @@
+export { ResourceTeamTable } from "./resourceTeamTable";
+export { ResourceTeamTableBody } from "./resourceTeamTableBody";
+export { ResourceTeamTableCell } from "./resourceTeamTableCell";
