+++ frontend/packages/app/src/app/pages/resource-management/team/components/table/resourceTeamTableCell.tsx
@@ -2,168 +2,21 @@
  * External dependencies.
  */
 import { useContext, useMemo } from "react";
-import { Table, TableBody } from "@next-pms/design-system/components";
 import { prettyDate } from "@next-pms/design-system/date";
-import {
-  ResourceTableCell,
-  ResourceTableRow,
-  ResourceTableHeader as ResourceTeamTableHeader,
-} from "@next-pms/resource-management/components";
-import { TableContextProvider } from "@next-pms/resource-management/store";
+import { ResourceTableCell } from "@next-pms/resource-management/components";
 import { getTableCellClass, getTodayDateCellClass, getCellBackGroundColor } from "@next-pms/resource-management/utils";
-import { InfiniteScroll } from "@/app/components/infiniteScroll";
 
 /**
  * Internal dependencies.
  */
 import { mergeClassNames } from "@/lib/utils";
-import type { ResourceAllocationObjectProps, ResourceAllocationProps } from "@/types/resource-management";
-import { ResourceExpandView } from "./expandView";
-import { EmptyTableBody, EmptyTableCell } from "../../components/empty";
-import { ResourceAllocationList } from "../../components/resourceAllocationList";
-import { ResourceFormContext } from "../../store/resourceFormContext";
-import { defaultEmployeeDayData, TeamContext } from "../../store/teamContext";
-import type {
-  AllocationDataProps,
-  DateProps,
-  EmployeeAllWeekDataProps,
-  EmployeeDataProps,
-  EmployeeResourceProps,
-} from "../../store/types";
-import { getIsBillableValue } from "../../utils/helper";
-
-/**
- * This component is responsible for loading the table for table view.
- *
- * @param props.onSubmit The on submit function used to handle soft update of allocation data.
- * @returns React.FC
- */
-const ResourceTeamTable = ({
-  onSubmit,
-  cellHeaderRef,
-  dateToAddHeaderRef,
-  handleVerticalLoadMore,
-}: {
-  onSubmit: (oldData: AllocationDataProps, data: AllocationDataProps) => void;
-  cellHeaderRef: React.LegacyRef<HTMLTableCellElement>;
-  dateToAddHeaderRef: string;
-  handleVerticalLoadMore: () => void;
-}) => {
-  const { teamData, apiController, getHasMore } = useContext(TeamContext);
-
-  const dates: DateProps[] = teamData.dates;
-  const isLoading = apiController.isLoading;
-  const hasMore = getHasMore();
-
-  if (dates.length == 0) {
-    return (
-      <div className="flex justify-center items-center h-full">
-        <EmptyTableBody />
-      </div>
-    );
-  }
-
-  return (
-    <TableContextProvider>
-      <Table className="relative">
-        <ResourceTeamTableHeader
-          dateToAddHeaderRef={dateToAddHeaderRef}
-          cellHeaderRef={cellHeaderRef}
-          dates={dates}
-          title="Members"
-        />
-        <InfiniteScroll isLoading={isLoading} hasMore={hasMore} verticalLodMore={handleVerticalLoadMore}>
-          <ResourceTeamTableBody onSubmit={onSubmit} />
-        </InfiniteScroll>
-      </Table>
-    </TableContextProvider>
-  );
-};
-
-/**
- * This function is responsible for rendering the table body for table view.
- *
- * @param props.onSubmit The on submit function used to handle soft update of allocation data.
- * @returns React.FC
- */
-const ResourceTeamTableBody = ({
-  onSubmit,
-}: {
-  onSubmit: (oldData: AllocationDataProps, data: AllocationDataProps) => void;
-}) => {
-  const { teamData } = useContext(TeamContext);
-
-  const data = teamData.data;
-  const dates = teamData.dates;
-
-  if (data.length == 0) {
-    return <EmptyTableBody />;
-  }
-
-  return (
-    <TableBody>
-      {data.map((employeeData: EmployeeDataProps) => {
-        return (
-          <ResourceTableRow
-            key={employeeData.name}
-            name={employeeData.name}
-            avatar={employeeData.image}
-            avatar_abbr={employeeData.employee_name}
-            avatar_name={employeeData.employee_name}
-            RowComponent={() => {
-              return (
-                <>
-                  {dates.map((week: DateProps, week_index: number) => {
-                    return week.dates.map((date: string, index: number) => {
-                      let employeeSingleDay = defaultEmployeeDayData;
-
-                      if (date in employeeData.all_dates_data) {
-                        employeeSingleDay = employeeData.all_dates_data[date];
-                      } else {
-                        employeeSingleDay = {
-                          ...employeeSingleDay,
-                          date: date,
-                          total_working_hours: employeeData.employee_daily_working_hours,
-                        };
-                      }
-
-                      let weekData = {
-                        total_allocated_hours: 0,
-                        total_working_hours: employeeData.employee_daily_working_hours * 5,
-                        total_worked_hours: 0,
-                      };
-
-                      if (week.key in employeeData.all_week_data) {
-                        weekData = employeeData.all_week_data[week.key];
-                      }
-
-                      return (
-                        <ResourceTeamTableCell
-                          key={`${employeeSingleDay.total_allocated_hours}-id-${Math.random()}`}
-                          employeeSingleDay={employeeSingleDay}
-                          weekData={weekData}
-                          employee={employeeData.name}
-                          employee_name={employeeData.employee_name}
-                          rowCount={index}
-                          midIndex={week_index}
-                          employeeAllocations={employeeData.employee_allocations}
-                          onSubmit={onSubmit}
-                        />
-                      );
-                    });
-                  })}
-                </>
-              );
-            }}
-            RowExpandView={() => {
-              return <ResourceExpandView employeeData={employeeData} onSubmit={onSubmit} />;
-            }}
-          />
-        );
-      })}
-    </TableBody>
-  );
-};
+import type { ResourceAllocationObjectProps, ResourceAllocationProps } from "@/types/resource_management";
+import { EmptyTableCell } from "../../../components/empty";
+import { ResourceAllocationList } from "../../../components/resource-allocation-list/resourceAllocationList";
+import { ResourceFormContext } from "../../../store/resourceFormContext";
+import { TeamContext } from "../../../store/teamContext";
+import type { AllocationDataProps, EmployeeAllWeekDataProps, EmployeeResourceProps } from "../../../store/types";
+import { getIsBillableValue } from "../../../utils/helper";
 
 /**
  * This component is responsible for loading the single cell of table view.
@@ -330,4 +183,4 @@ const ResourceTeamTableCell = ({
   );
 };
 
-export { ResourceTeamTable };
+export { ResourceTeamTableCell };
