+++ frontend/packages/app/src/app/pages/resource-management/timeline/components/header/index.ts
@@ -1,0 +1,3 @@
+export { ResourceTimLineHeaderSection } from "./resourceTimLineHeaderSection";
+export { TimeLineDateHeader } from "./timeLineDateHeader";
+export { TimeLineIntervalHeader } from "./timeLineIntervalHeader";
