+++ frontend/packages/app/src/app/pages/resource-management/timeline/components/item/types.ts
@@ -1,0 +1,7 @@
+interface ItemAllocationActionDialogProps {
+  handleMove: () => void;
+  handleCopy: () => void;
+  handleCancel: () => void;
+}
+
+export type { ItemAllocationActionDialogProps };
