+++ frontend/packages/resource-management/src/components/table/row/index.ts
@@ -1,0 +1,2 @@
+export { ResourceTableRow } from "./resourceTableRow";
+export { TableDisabledRow } from "./tableDisabledRow";
